Authentication
The server now requires authentication for everything. The client now has a structure for an authenticated get, this can be extrapolated to other methods if need be. There is also a basic login for the GUI. This is not hooked up to the server, but will be once a database is implemented.